primitives: use alloy `Header` struct (#10691)

+++ crates/exex/exex/src/backfill/job.rs
@@ -125,7 +125,7 @@ where
             if self.thresholds.is_end_of_batch(
                 block_number - *self.range.start(),
                 bundle_size_hint,
-                cumulative_gas,
+                cumulative_gas as u64,
                 batch_start.elapsed(),
             ) {
                 break
@@ -138,7 +138,7 @@ where
             range = ?*self.range.start()..=last_block_number,
             block_fetch = ?fetch_block_duration,
             execution = ?execution_duration,
-            throughput = format_gas_throughput(cumulative_gas, execution_duration),
+            throughput = format_gas_throughput(cumulative_gas as u64, execution_duration),
             "Finished executing block range"
         );
         self.range = last_block_number + 1..=*self.range.end();
